Week_5/Day_5/ExercisesXP/requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.1.0
phonenumbers==8.13.10
psycopg2==2.9.6
sqlparse==0.4.4
tzdata==2023.3
